+++ demos/dra-slinky/demo.sh
@@ -17,7 +17,7 @@ step "make build TARGETS=kwok-nodes"
 
 step "delete_cluster"
 
-step "./scripts/create-test-cluster.sh -m ./tests/models/medium.yaml"
+step "./scripts/create-test-cluster.sh -m ./tests/models/large.yaml"
 
 step "./demos/dra-slinky/update-labels.sh"
 
@@ -27,4 +27,4 @@ step "deploy_topograph demos/dra-slinky/values.dra-slinky.kwok.yaml"
 
 # step "kubectl --context \"$KUBE_CONTEXT\" -n topograph logs -l app.kubernetes.io/name=topograph"
 
-step "kubectl --context \"$KUBE_CONTEXT\" -n slurm get cm slurm-config-extra -o jsonpath='{.data.topology\.conf}'"
+step "kubectl --context \"$KUBE_CONTEXT\" -n slurm get cm slurm-config-extra -o jsonpath='{.data.topology\.conf}' | grep -v '#'"